covidmeter/backend/covidmeter/api/user/tests/test_views/test_user_list_create_view/setup.py
from rest_framework.test import APITransactionTestCase

from api.user.tests.factories import UserFactory


class SetUpClass(APITransactionTestCase):
    reset_sequences = True

    def setUp(self) -> None:
        self.users = UserFactory.create_batch(20)
