covidmeter/backend/covidmeter/requirements.txt
asgiref==3.4.1
autopep8==1.5.7
black==21.9b0
click==8.0.1
Django==3.2.6
django-cleanup==5.2.0
django-cors-headers==3.8.0
django-debug-toolbar==3.2.2
django-environ==0.5.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
factory-boy==3.2.0
Faker==8.12.1
Markdown==3.3.4
mypy-extensions==0.4.3
pathspec==0.9.0
pep8==1.7.1
Pillow==8.3.2
platformdirs==2.3.0
psycopg2-binary==2.9.1
pycodestyle==2.7.0
PyJWT==2.1.0
python-dateutil==2.8.2
pytz==2021.1
regex==2021.8.28
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.2
